make paint texture opacity level use integer

+++ src/noggit/MapChunk.cpp
@@ -885,7 +885,7 @@ void MapChunk::switchTexture(scoped_blp_texture_reference oldTexture, scoped_blp
   texture_set->switchTexture(oldTexture, newTexture);
 }
 
-bool MapChunk::paintTexture(math::vector_3d const& pos, Brush* brush, float strength, float pressure, scoped_blp_texture_reference texture)
+bool MapChunk::paintTexture(math::vector_3d const& pos, Brush* brush, uint strength, float pressure, scoped_blp_texture_reference texture)
 {
   return texture_set->paintTexture(xbase, zbase, pos.x, pos.z, brush, strength, pressure, texture);
 }
